[DI] move phpdoc node decorators to NodeDecorator namespace, inject explicitly in order
Move the four PhpDocNodeDecorator services out of the PhpDocParser
namespace into a dedicated NodeDecorator namespace, and inject them into
BetterPhpDocParser as explicit constructor params instead of a tagged
collection.

The decorators have a required run order: DoctrineAnnotationDecorator
must run before ArrayItemClassNameDecorator, which resolves class names
in the array items the former produces (e.g. inside doctrine
annotations). Explicit ordered injection guarantees this; tagged or
autodiscovered registration would not.
