+++ rules-tests/DeadCode/Rector/Assign/RemoveDoubleSelfAssignRector/Fixture/skip_different_variable.php.inc
@@ -2,4 +2,4 @@
 
 namespace Rector\Tests\DeadCode\Rector\Assign\RemoveDoubleSelfAssignRector\Fixture;
 
-$first = $second = createValidator();
+$first = $second = 1234;
