+++ src/BetterPhpDocParser/NodeDecorator/ConstExprClassNameDecorator.php
@@ -2,22 +2,22 @@
 
 declare(strict_types=1);
 
-namespace Rector\BetterPhpDocParser\PhpDocParser;
+namespace Rector\BetterPhpDocParser\NodeDecorator;
 
 use PhpParser\Node as PhpNode;
+use PHPStan\PhpDocParser\Ast\ConstExpr\ConstFetchNode;
 use PHPStan\PhpDocParser\Ast\Node;
 use PHPStan\PhpDocParser\Ast\PhpDoc\PhpDocNode;
 use Rector\BetterPhpDocParser\Contract\PhpDocParser\PhpDocNodeDecoratorInterface;
-use Rector\BetterPhpDocParser\PhpDoc\ArrayItemNode;
 use Rector\BetterPhpDocParser\ValueObject\PhpDocAttributeKey;
 use Rector\PhpDocParser\PhpDocParser\PhpDocNodeTraverser;
 use Rector\StaticTypeMapper\Naming\NameScopeFactory;
 
 /**
- * Decorate node with fully qualified class name for annotation:
- * e.g. @ORM\Column(type=Types::STRING, length=100, nullable=false)
+ * Decorate node with fully qualified class name for const epxr,
+ * e.g. Direction::*
  */
-final readonly class ArrayItemClassNameDecorator implements PhpDocNodeDecoratorInterface
+final readonly class ConstExprClassNameDecorator implements PhpDocNodeDecoratorInterface
 {
     public function __construct(
         private NameScopeFactory $nameScopeFactory,
@@ -35,29 +35,20 @@ final readonly class ArrayItemClassNameDecorator implements PhpDocNodeDecoratorI
         $this->phpDocNodeTraverser->traverseWithCallable($phpDocNode, '', function (Node $node) use (
             $phpNode
         ): Node|null {
-            if (! $node instanceof ArrayItemNode) {
+            if (! $node instanceof ConstFetchNode) {
                 return null;
             }
 
-            if (! is_string($node->value)) {
-                return null;
-            }
-
-            $splitScopeResolution = explode('::', $node->value);
-            if (count($splitScopeResolution) !== 2) {
-                return null;
-            }
-
-            $className = $this->resolveFullyQualifiedClass($splitScopeResolution[0], $phpNode);
+            $className = $this->resolveFullyQualifiedClass($node, $phpNode);
             $node->setAttribute(PhpDocAttributeKey::RESOLVED_CLASS, $className);
 
             return $node;
         });
     }
 
-    private function resolveFullyQualifiedClass(string $className, PhpNode $phpNode): string
+    private function resolveFullyQualifiedClass(ConstFetchNode $constFetchNode, PhpNode $phpNode): string
     {
         $nameScope = $this->nameScopeFactory->createNameScopeFromNodeWithoutTemplateTypes($phpNode);
-        return $nameScope->resolveStringName($className);
+        return $nameScope->resolveStringName($constFetchNode->className);
     }
 }
